Add homepage and adjusted route
